requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.11.28
cloudpickle==1.2.2
cvxpy==1.0.25
cycler==0.10.0
cytoolz==0.10.1
dask==2.9.1
decorator==4.4.1
dill==0.3.1.1
ecos==2.0.7.post1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.3
imageio==2.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
multiprocess==0.70.9
networkx==2.4
nltk==3.4.5
numpy==1.18.1
olefile==0.46
osqp==0.6.1
pandas==0.25.3
Pillow==7.0.0
progress-bar==8
protobuf==3.10.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.15.0
scipy==1.3.2
scs==2.1.1.post2
sip==4.19.13
six==1.13.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
Werkzeug==0.16.0
wrapt==1.11.2
xgboost==0.90
